[*] разнёс функции из assember.cpp в make_cmd.cpp и program.cpp
